dataset_creation/requirements.txt
numpy
Pillow
tqdm
cv2
nuscenes-devkit
descartes
matplotlib
pyquaternion
shapely
